[S07P12B102-267] Feat. 개발 - 피드: [상세정보] 커스텀 칵테일 ID를 확인하고 해당 ID를 가지는 커스텀 칵테일 정보 API 구현

+++ backend/src/main/java/com/jourgeois/backend/repository/PostBookmarkRepository.java
@@ -10,6 +10,6 @@ import java.util.List;
 import java.util.Optional;
 
 public interface PostBookmarkRepository extends JpaRepository<PostBookmark, PostBookmarkId> {
-    Long countByPostId(Post p_id);
+    Integer countByPostId(Post p_id);
     List<PostBookmark> findByPostId(Post p_id, Pageable pageable);
 }
